# docs/swagger.yaml
basePath: /api
definitions:
  internal_api.BadgeResponse:
    properties:
      badge:
        $ref: '#/definitions/verve_internal_models.Badge'
      rules:
        items:
          $ref: '#/definitions/verve_internal_models.AchievementRule'
        type: array
    type: object
  internal_api.CreateBadgeRequest:
    properties:
      description:
        example: Awarded to users who complete all achievements
        type: string
      icon_url:
        example: https://example.com/badges/master.png
        type: string
      name:
        example: Achievement Master
        type: string
      points:
        example: 100
        type: integer
      rules:
        items:
          $ref: '#/definitions/verve_internal_models.AchievementRule'
        type: array
    required:
    - description
    - name
    type: object
...
  internal_api.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
    type: object
...
  internal_api.SuccessResponse:
    properties:
      message:
        example: Operation successful
        type: string
    type: object
...
  internal_api.UpdateBadgeRequest:
    properties:
      description:
        example: Awarded to users who complete all achievements
        type: string
      icon_url:
        example: https://example.com/badges/master.png
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Achievement Master
        type: string
      points:
        example: 100
        type: integer
    type: object
  internal_api.UpdateUserRequest:
    properties:
      display_name:
        example: John Doe
        type: string
      pin_required_for_transfer:
        example: true
        type: boolean
      profile_photo_url:
        example: https://example.com/photo.jpg
        type: string
    type: object
  verve_internal_models.AchievementRule:
    properties:
      badge_id:
        type: integer
      condition_value:
        items:
          type: integer
        type: array
      created_at:
        type: string
      created_by:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      rule_type:
        type: string
    type: object
  verve_internal_models.Badge:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      icon_url:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      points:
        type: integer
    type: object
...
  verve_internal_models.User:
    description: User information including display name and profile settings
    properties:
      created_at:
        type: string
      display_name:
        example: John Doe
        type: string
      email:
        example: john.doe@example.com
        type: string
      id:
        example: 1
        type: integer
      pin_required_for_transfer:
        example: true
        type: boolean
      profile_photo_url:
        example: https://example.com/photo.jpg
        type: string
      provider:
        description: OAuth provider (google, okta, or local)
        example: google
        type: string
      provider_user_id:
        description: ID from the OAuth provider
        type: string
      roles:
        example:
        - '[''user'''
        - '''admin'']'
        items:
          type: string
        type: array
      updated_at:
        type: string
      username:
        example: john.doe
        type: string
    type: object
  verve_internal_models.UserBadge:
    properties:
      awarded_at:
        type: string
      awarded_by:
        description: Pointer to allow NULL for system-awarded badges
        type: integer
      badge_id:
        type: integer
      id:
        type: integer
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: Verve API
  version: "1.0"
paths:
...
  /badges:
    get:
      description: Get a list of all badges (admins see all, users see only active)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/verve_internal_models.Badge'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List all badges
      tags:
      - badges
    post:
      consumes:
      - application/json
      description: Create a new badge with optional achievement rules
      parameters:
      - description: Badge details
        in: body
        name: badge
        required: true
        schema:
          $ref: '#/definitions/internal_api.CreateBadgeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created badge
          schema:
            $ref: '#/definitions/verve_internal_models.Badge'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new badge
      tags:
      - badges
  /badges/{id}:
    get:
      description: Get a badge's details and achievement rules
      parameters:
      - description: Badge ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.BadgeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: Badge not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a badge
      tags:
      - badges
    put:
      consumes:
      - application/json
      description: Update an existing badge's details
      parameters:
      - description: Badge ID
        in: path
        name: id
        required: true
        type: integer
      - description: Badge details to update
        in: body
        name: badge
        required: true
        schema:
          $ref: '#/definitions/internal_api.UpdateBadgeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/verve_internal_models.Badge'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: Badge not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a badge
      tags:
      - badges
  /badges/{id}/award/{user_id}:
    post:
      description: Award a badge to a specific user
      parameters:
      - description: Badge ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.SuccessResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: Badge or user not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Award badge to user
      tags:
      - badges
  /badges/{id}/holders:
    get:
      description: Get all users who have been awarded a specific badge
      parameters:
      - description: Badge ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/verve_internal_models.UserBadge'
            type: array
        "400":
          description: Invalid badge ID
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: Badge not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get badge holders
      tags:
      - badges
...
  /user/{id}:
    get:
      description: Get details of a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/verve_internal_models.User'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user details
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the authenticated user's profile information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User details to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_api.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/verve_internal_models.User'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Forbidden - Can only update own profile
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user profile
      tags:
      - users
  /user/{id}/badges:
    get:
      description: Get all badges that have been awarded to the authenticated user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/verve_internal_models.UserBadge'
            type: array
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Forbidden - Can only view your own badges
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user's badges
      tags:
      - badges
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"